CommUnity/requirements.txt
cryptography==44.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.6
django-allauth==0.61.0
et_xmlfile==2.0.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
oauthlib==3.2.2
openpyxl==3.1.5
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
tzdata==2025.1
google-auth google-auth-oauthlib google-auth-httplib2 google-api-python-client